Pipeline refactorings
Renamed SignalAnalyser to Conductor.

Conductor is initialized with the MadSchMIDI2PCM. It is the single point of communication.
The Conductor is passed into every pipeline. The pipeline, reports to back to the Conductor. Conductor forwards messages from the pipelines to other pipelines if necessary.

Conductor holds preferences of the detectors.